Additional basic matrix operation support
Implement a few additional basic operations for matrices. Also reverse
template parameter order and rename aliases. Replace [] operator with
() operator to enable direct element access for matrices. Restrict
matrix dimensions to N,M = 2, 3, or 4 to reduce complexity for the time
being - it's unlikely that we'll need to support higher dimensions.