Add target hit detection example
Add a simple example demonstrating using Vec to determine if a shot from
a particular position/direction will hit a circular target with a
specified position, norma, and radius.

Also update the ostream operator to add an extra space between vector
elements. This will hopefully make the output a little easier to read.