Add affine transform class template and unit tests
Add a new header file with an AffineTransform class template. This class
extends the matrix class with a template size of M+1 to enable a
homogeneous coordinate representation. Provides convenience constructors
and accessors for the linear transformation matrix and translation
vector components of the transform.

Add unit tests for the new class template.